- Renaming package to servicebus, from sb - Removing the mis-cased readme.md, in favor of README.md - Updated the readme to accomodate the new folder structure, and to provide a pointer to the package itself, just in case someone finds it.
