+++ sdk/messaging/stress/internal/servicebus/tests/constant_detachment.go
@@ -12,7 +12,7 @@ import (
 	"sync/atomic"
 	"time"
 
-	"stress/internal/sb/shared"
+	"stress/internal/servicebus/shared"
 
 	"github.com/Azure/azure-sdk-for-go/sdk/messaging/azservicebus"
 	"github.com/Azure/azure-sdk-for-go/sdk/messaging/azservicebus/admin"
