+++ sdk/messaging/stress/internal/servicebus/tests/constant_detachment_sender.go
@@ -10,7 +10,7 @@ import (
 	"sync"
 	"time"
 
-	"stress/internal/sb/shared"
+	"stress/internal/servicebus/shared"
 
 	"github.com/Azure/azure-sdk-for-go/sdk/messaging/azservicebus"
 	"github.com/Azure/azure-sdk-for-go/sdk/messaging/azservicebus/admin"
