+++ sdk/messaging/stress/internal/servicebus/tests/empty_sessions.go
@@ -12,7 +12,7 @@ import (
 	"strings"
 	"time"
 
-	"stress/internal/sb/shared"
+	"stress/internal/servicebus/shared"
 
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/to"
 	"github.com/Azure/azure-sdk-for-go/sdk/messaging/azservicebus"
