+++ sdk/messaging/stress/internal/servicebus/tests/idle_fast_reconnect.go
@@ -9,7 +9,7 @@ import (
 	"log"
 	"time"
 
-	"stress/internal/sb/shared"
+	"stress/internal/servicebus/shared"
 
 	"github.com/Azure/azure-sdk-for-go/sdk/messaging/azservicebus"
 )
