+++ sdk/messaging/stress/internal/servicebus/tests/infinite_send_and_receive.go
@@ -10,7 +10,7 @@ import (
 	"log"
 	"time"
 
-	"stress/internal/sb/shared"
+	"stress/internal/servicebus/shared"
 
 	"github.com/Azure/azure-sdk-for-go/sdk/messaging/azservicebus"
 )
