+++ sdk/messaging/stress/internal/servicebus/tests/mostly_idle_receiver.go
@@ -9,7 +9,7 @@ import (
 	"sync"
 	"time"
 
-	"stress/internal/sb/shared"
+	"stress/internal/servicebus/shared"
 
 	"github.com/Azure/azure-sdk-for-go/sdk/messaging/azservicebus"
 )
