+++ sdk/messaging/stress/internal/servicebus/tests/open_close_measurements.go
@@ -12,7 +12,7 @@ import (
 	"sync"
 	"time"
 
-	"stress/internal/sb/shared"
+	"stress/internal/servicebus/shared"
 
 	"github.com/Azure/azure-sdk-for-go/sdk/messaging/azservicebus"
 	"github.com/Azure/azure-sdk-for-go/sdk/messaging/azservicebus/admin"
