+++ sdk/messaging/stress/internal/servicebus/tests/receive_cancellation.go
@@ -9,7 +9,7 @@ import (
 	"fmt"
 	"time"
 
-	"stress/internal/sb/shared"
+	"stress/internal/servicebus/shared"
 
 	"github.com/Azure/azure-sdk-for-go/sdk/messaging/azservicebus"
 )
